minified/agui-agent-redis/pyproject.toml
[project]
name = "magnum-opus-agui-agent"
version = "0.1.0"
description = "AG-UI gateway and ADK-backed Agent Service with Redis thread metadata."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ag-ui-adk>=0.6.1",
    "anthropic>=0.43.0",
    "fastapi>=0.115.0",
    "google-adk>=1.32.0",
    "httpx>=0.27.0",
    "pydantic>=2.8.0",
    "redis>=5.0.0",
    "uvicorn[standard]>=0.30.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["setuptools>=70.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = [
    "services/ag_ui_gateway",
    "services/agent_service",
]
include = [
    "gateway_app*",
    "adk_agent_service*",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = [
    "services/ag_ui_gateway",
    "services/agent_service",
]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["ASYNC", "B", "E", "F", "I", "UP"]
